Restructures dial and group directories
