+++ src/Group/GroupDetails/GroupDetails.css
